requirements.txt
click
sqlalchemy
python-magic
git_annex_adapter
